@@ -10,7 +10,7 @@ homepage = "https://www.nushell.sh"
 license = "MIT"
 name = "nu"
 repository = "https://github.com/nushell/nushell"
-rust-version = "1.86.0"
+rust-version = "1.87.0"
 version = "0.106.2"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -72,6 +72,7 @@ brotli = "7.0"
 byteorder = "1.5"
 bytes = "1"
 bytesize = "1.3.3"
+byteyarn = "0.5"
 calamine = "0.28"
 chardetng = "0.1.17"
 chrono = { default-features = false, version = "0.4.34" }
@@ -100,6 +101,7 @@ indicatif = "0.17"
 interprocess = "2.2.0"
 is_executable = "1.0"
 itertools = "0.14"
+lean_string = { version = "0.5", features = ["serde"] }
 libc = "0.2"
 libproc = "0.14"
 log = "0.4"
@@ -142,7 +144,7 @@ rand = "0.9"
 getrandom = "0.2" # pick same version that rand requires
 rand_chacha = "0.9"
 ratatui = "0.29"
-rayon = "1.10"
+rayon = "1.11"
 reedline = "0.41.0"
 rmp = "0.8"
 rmp-serde = "1.3"
@@ -203,6 +205,7 @@ webpki-roots = "1.0"
 # todo = "warn"
 unchecked_duration_subtraction = "warn"
 used_underscore_binding = "warn"
+result_large_err = "allow"
 
 [lints]
 workspace = true
